+++ bin/cibw/pre_build.sh
@@ -1,0 +1,23 @@
+#!/bin/bash
+
+# Copyright (C) 2024 qBraid
+#
+# This file is part of pyqasm
+#
+# Pyqasm is free software released under the GNU General Public License v3
+# or later. You can redistribute and/or modify it under the terms of the GPL v3.
+# See the LICENSE file in the project root or <https://www.gnu.org/licenses/gpl-3.0.html>.
+#
+# THERE IS NO WARRANTY for pyqasm, as per Section 15 of the GPL v3.
+
+set -e 
+
+echo "Running pre_build.sh"
+
+# Upgrade pip
+python -m pip install --upgrade pip
+
+# Install required packages
+pip install setuptools wheel cython
+
+echo "Finished running pre_build.sh"