+++ bin/install_wheel_extras.sh
@@ -17,8 +17,9 @@
 # Extras are specified with --extra flags.
 #
 # Example Usage:
-#   ./install_wheel_extras.sh ./dist --extra extra1 --extra extra2
-#   ./install_wheel_extras.sh ./dist
+#  ./install_wheel_extras.sh ./dist
+#   ./install_wheel_extras.sh ./dist --type wheel --extra extra1 --extra extra2
+#   ./install_wheel_extras.sh ./dist --type sdist --extra extra1 --extra extra2
 ################################################################################
 
 # Check if at least one argument (the directory path) is provided
@@ -30,6 +31,17 @@ fi
 # The first argument is the directory path
 DIST=$1
 shift # Remove the directory path from the arguments list
+
+# The second argument(optional) is the type of file to install (wheel or sdist)
+TYPE="wheel"
+if [ "$1" == "--type" ]; then
+    TYPE=$2
+    if [ "$TYPE" != "wheel" ] && [ "$TYPE" != "sdist" ]; then
+        echo "Invalid type: $TYPE"
+        exit 1
+    fi
+    shift 2
+fi
 
 # Initialize extras array
 EXTRAS=()
@@ -51,21 +63,27 @@ done
 # Combine extras into a comma-separated string
 EXTRAS_STR=$(IFS=, ; echo "${EXTRAS[*]}")
 
-# Use ls and grep to find the first .whl file
-WHEEL_FILE=$(ls $DIST | grep '\.whl$' | head -n 1)
-
-if [ -z "$WHEEL_FILE" ]; then
-    echo "No .whl file found in the specified directory."
-    exit 1
+# Use ls and grep to find the first installable
+if [ "$TYPE" == "sdist" ]; then
+    INSTALL_FILE=$(ls $DIST | grep '\.tar\.gz$' | head -n 1)
 else
-    echo "Wheel file found: $WHEEL_FILE"
+    INSTALL_FILE=$(ls $DIST | grep '\.whl$' | head -n 1)
+fi
+
+if [ -z "$INSTALL_FILE"]; then
+    if [ "$TYPE" == "sdist" ]; then
+        echo "No .tar.gz file found in $DIST"
+    else
+        echo "No .whl file found in $DIST"
+    fi
+    exit 1
 fi
 
 # Build the pip install command with extras, if provided
 if [ -n "$EXTRAS_STR" ]; then
-    INSTALL_COMMAND="python3 -m pip install '$DIST/$WHEEL_FILE[$EXTRAS_STR]'"
+    INSTALL_COMMAND="python3 -m pip install '$DIST/$INSTALL_FILE[$EXTRAS_STR]'"
 else
-    INSTALL_COMMAND="python3 -m pip install '$DIST/$WHEEL_FILE'"
+    INSTALL_COMMAND="python3 -m pip install '$DIST/$INSTALL_FILE'"
 fi
 
 # Execute the pip install command