+++ CHANGELOG.md
@@ -105,6 +105,7 @@ Success: no issues found in 2 source files
     - Migrated `pyqasm.linalg._kronecker_factor` to `pyqasm.linalg_cy` with ~60% speedup
     - Migrated `pyqasm.linalg._so4_to_so2()` to to `pyqasm.linalg_cy` with ~5% speedup
 - Changed source code directory from `./pyqasm` to `./src/pyqasm` to prevents conflicts between the local source directory and the installed package in site-packages, ensuring Python's module resolution prioritizes the correct version. Required for local testing with new Cython build step ([#83](https://github.com/qBraid/pyqasm/pull/83))
+- Updated the build process for `pyqasm` due to Cython integration. Wheels are now built for each platform and uploaded to PyPI along with the source distributions ([#88](https://github.com/qBraid/pyqasm/pull/88))
 
 ### Deprecated